Revert "float32 is big enough for temperatures"
This reverts commit 9ce8aa1227cb8484bf23af63c45b9031fee8be66.

+++ 2014-01/celsius/convert/first.go
@@ -5,13 +5,13 @@ import (
 )
 
 func main() {
-	var c, f float32
+	var c, f float64
 	c = -1.0
 	f = CelsiusToFahrenheit(c)
 	fmt.Printf("%4.2fºC is %4.2fºF\n", c, f)
 }
 
 // CelsiusToFahrenheit converts temperature to share with our American friends.
-func CelsiusToFahrenheit(c float32) float32 {
+func CelsiusToFahrenheit(c float64) float64 {
 	return (c * 9 / 5) + 32
 }
